# Scrabble-BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api pour scrabble",
        "description": "Implementation de Swagger avec Laravel9 pour documneter l'api",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/v1/joueur/{idJoueur}": {
            "get": {
                "tags": [
                    "joueurs"
                ],
                "summary": "Trouver un joueurs a partie son id ",
                "description": "Trouver un joueurs a partie son id",
                "operationId": "getJoueur",
                "parameters": [
                    {
                        "name": "idJoueur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/inscrire": {
            "post": {
                "tags": [
                    "joueur"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "c44c23b493b72b5a17859078445ffc47",
                "requestBody": {
                    "description": "Station to create",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Returns the created station"
                    }
                }
            }
        },
        "/v1/joueurs": {
            "get": {
                "tags": [
                    "joueurs"
                ],
                "summary": "la liste des joueurs",
                "description": "la liste des joueurs",
                "operationId": "getJoueurs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StationPostRequest": {
                "properties": {
                    "nom": {
                        "schema": "JoueurRequest",
                        "description": "joueur nom",
                        "type": "string"
                    },
                    "photo": {
                        "schema": "JoueurRequest",
                        "description": "The station title",
                        "type": "file"
                    },
                    "partie": {
                        "schema": "JoueurRequest",
                        "description": "The station title",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "joueurs": {
                "properties": {
                    "nom": {
                        "schema": "JoueurResource",
                        "description": "The station name",
                        "type": "string"
                    },
                    "photo": {
                        "schema": "JoueurResource",
                        "description": "The station title",
                        "type": "string"
                    },
                    "partie": {
                        "schema": "JoueurResource",
                        "description": "The station latitude",
                        "type": "number"
                    },
                    "chevalet": {
                        "schema": "JoueurResource",
                        "description": "The station longitude",
                        "type": "string"
                    },
                    "statutJoueur": {
                        "schema": "JoueurResource",
                        "description": "The station longitude",
                        "type": "string"
                    },
                    "ordre": {
                        "schema": "JoueurResource",
                        "description": "The station longitude",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}